Extract HtmlRenderer from HtmlGenerator
Split rendering logic into separate class:
- HtmlGenerator: data extraction and processing (13 methods)
- HtmlRenderer: HTML rendering and CSS (17 methods)
